requirements.txt
# Core dependencies
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Statistical analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Optional: for running notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
